Create Troubleshooting.md
The latest versions of MongoDB have caused this issue on certain CPUs. Downgrading to MongoDB version 4.2 is the solution to this problem. I’ve provided the steps to apply this fix, but I’m not sure how exactly to incorporate this into the documentation. However, you can see the solution in the document.